refactor(eve): derive the TUI title from the resolved target
Move resolveTuiTitle into target.ts and take the DevelopmentTuiTarget
union instead of loose {remoteServerUrl, appRoot} fields, so the
local/remote decision lives in one place. run.ts builds the target once
and passes it to both the title and the TUI input. Test relocated
alongside the function.
